+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - `vector` `0.26.0` -> `0.33.0` ([#361], [#377]).
 - `operator-rs` `0.44.0` -> `0.55.0` ([#360], [#376], [#377]).
 - jmx-exporter now referenced via soft link without version ([#377]).
+- Service discovery now exposes the cluster service to enable HA ([#382]).
 
 ### Removed
 
@@ -25,6 +26,7 @@ All notable changes to this project will be documented in this file.
 [#365]: https://github.com/stackabletech/hive-operator/pull/365
 [#376]: https://github.com/stackabletech/hive-operator/pull/376
 [#377]: https://github.com/stackabletech/hive-operator/pull/377
+[#382]: https://github.com/stackabletech/hive-operator/pull/382
 
 ## [23.7.0] - 2023-07-14